+++ R/das_chop_condition.R
@@ -15,7 +15,10 @@
 #'   Default is \code{NULL} since these distances should have already been calculated
 #' @param num.cores see \code{\link{das_effort}}
 #'
-#' @details This function is intended to be called by \code{\link{das_effort}}
+#' @details WARNING - do not call this function directly!
+#'   It is intended for internal package use only, and is exported for documentation purposes only.
+#'
+#'   This function is intended to be called by \code{\link{das_effort}}
 #'   when the "condition" method is specified.
 #'   Thus, \code{x} must be filtered for events (rows) where either
 #'   the 'OnEffort' column is \code{TRUE} or the 'Event' column is "E";
@@ -62,8 +65,6 @@
 #'   \item segdata: data frame with one row for each segment, and columns with
 #'     relevant data (see \code{\link{das_effort}} for specifics)
 #' }
-#'
-#' @keywords internal
 #'
 #' @export
 das_chop_condition <- function(x, ...) UseMethod("das_chop_condition")