+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added 'strata.files' argument to `das_effort` to allow users to split effort by strata
 
+* The 'segdata' output of `das_effort` now includes DateTime1 and DateTime2, the date/time of the start and end points of each segment. The column order of the segdata data frame was also slightly changed for consistency
+
 * Changed the name of the strata argument of `das_intersects_strata` to 'strata.files' for consistency
 
 * Fixed a bug in `das_check` where multiple files could not be checked at once
@@ -11,6 +13,8 @@
 * Fixed an issue where the sample stratum file did not overlap with any of the sample data
 
 * Added documentation describing how `das_sight` handles additional sperm whale group size estimates
+
+* `das_chop_condition`, `das_chop_equallength`, `das_chop_section`, and `das_segdata` are now exported for documentation purposes only
 
 
 # swfscDAS 0.4.0